requirements.txt
playwright
playwright-stealth
pytest
pytest-xdist
pyyaml
allure-pytest